emit certain events for unique tweets
Summary: This is my implementation for handling duplicate tweets. For site streams, and every stream created by the pool, it will keep track of the last 1000 tweet ids. When a new tweet comes in it determines if it's a duplicate or not by comparing the id.

Test Plan: npm test

Reviewers: andybons, joy, steve

Reviewed By: andybons

CC: Engineering

Differential Revision: https://phabricator.poptip.com/D379