use an event emitter instead of streamify to queue requests
Summary:
I released that the streamify module is not necessary for what it was being
used for in this situation. It is not important to immediately return a stream
object right away since the stream.js file takes care of all of that.

Test Plan: `npm test` on node v0.8.21 and v0.10.0 and it ran fine with flood.

Reviewers: andybons, joy, steve

Reviewed By: andybons

CC: Engineering

Differential Revision: https://phabricator.poptip.com/D546