dont keep requesting on disconnect
Summary:
Site stream use control streams to add/remove users and for status checks.
Whenever a site stream connects, it has a different control stream uri.
This patch makes it so queued requests for a control stream will not
go through if a stream has been disconnected. Instead, they will be queued
until the stream emits the new control uri.

This isn't meant to fix any current errors, but if there are any errors in the
future, it will let us find the source of them much easier. And will not
bother Twitter as much with requests which will return with a 404.

Test Plan: npm test and ran it on flood

Reviewers: andybons, joy, steve

Reviewed By: andybons

CC: Engineering

Differential Revision: https://phabricator.poptip.com/D575