use a GET request instead
Summary:
this is an attempt to minimize the errors shown in site status
GET requests weren't supported by streams at first,
you could only open them with POST.
now all streams support it.

This attempt is to see if for some reason request (http module we use)
or Twitter is parsing the POST body incorrectly. It's usually easier
to parse a URL.

Test Plan: tests and TPM comparison

Reviewers: joy

Reviewed By: joy

Differential Revision: https://phabricator.poptip.com/D2651